load-graph: Change variable from double to int
While used to calculate floats, on its own the values are integers.

+++ src/load-graph.h
@@ -63,7 +63,7 @@ struct LoadGraph
   guint frames_per_unit;
   guint graph_dely;
   guint real_draw_height;
-  double graph_delx;
+  guint graph_delx;
   guint graph_buffer_offset;
 
   std::vector<GdkRGBA> colors;
